requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h2bbff1b_0
ca-certificates=2022.4.26=haa95532_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
lz4-c=1.9.3=h2bbff1b_1
matplotlib-base=3.5.1=py310hd77b12b_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl-service=2.4.0=py310h2bbff1b_0
nest-asyncio=1.5.5=py310haa95532_0
numpy-base=1.21.5=py310h206c741_3
pandas-datareader=0.10.0=pyhd3eb1b0_0
pandas-profiling=2.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
scikit-learn=1.0.2=py310h4ed8f06_0
tangled-up-in-unicode=0.1.0=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
